Dts downmix: Renamed gstdtsdec plugin to gstdtsdownmix.
 This chnage is to avoid conflicts with the existing bad plugin gstdtsdec.

	modified:   Makefile.am
	renamed:    gstdtsdec.c -> gstdtsdownmix.c
	renamed:    gstdtsdec.h -> gstdtsdownmix.h
